test slice, move function
